refactor folder and module structure
